Changing API for converting Excel numbers/dates back and forth
After adding some tests and trying to use with formats module, it seems that
the previous API was "not ideal." We now always convert Excel numbers to a full
NaiveDateTime object. It will then be up to the caller to determine if they
only want to use part of that value (e.g., if the date == 0, assume it is a
time object).

Since we limit the output to dates or numbers, it became obvious that we should
be using a `Result` instead of a custom struct for the return value. If it
returns a number, we essentially hit a snag / ran into an error.

Also added a bunch of tests (going from both number to date and vice versa).